Lamnda Functions/Client/ChatCrew/requirements.txt
# CrewAI framework
crewai~=0.150.0

# LLM management
langchain

# Gemini API key management
python-dotenv==1.1.1

# HTTP requests to external services (like product APIs and store-location APIs)
httpx==0.27.0

# AWS SDK for Python (DynamoDB interaction)
boto3~=1.39.14

# Data validation and schema for tool inputs
pydantic==2.11.7

# Command-line argument parsing (used in chat_simulator.py)
argparse~=1.4.0

# For compatibility with older ast.literal_eval-based workflows
#astunparse; python_version < "3.9"

gradio~=5.44.1
requests~=2.32.4

pyarrow==15.0.2
numpy==1.26.4
pandas==2.2.2